do not let the user send an empty MAC

+++ html/captive-portal/lib/captiveportal/PacketFence/Controller/DeviceRegistration.pm
@@ -52,21 +52,23 @@ sub index : Path : Args(0) {
     } elsif ( $request->param('cancel') ) {
         $c->user_session({});
         $c->detach('login');
-    } elsif ( $request->param('device_mac') ) {
-        # User is authenticated and requesting to register a device
-        my $device_mac = clean_mac($request->param('device_mac'));
-        my $device_type;
-        $device_type = $request->param('console_type') if ( defined($request->param('console_type')) );
-        if(valid_mac($device_mac)) {
-            # Register device
-            $c->forward('registerNode', [ $pid, $device_mac, $device_type ]);
-            unless ($c->has_errors) {
-                $c->stash(status_msg  => [ "The MAC address %s has been successfully registered.", $device_mac ]);
-                $c->detach('landing');
+    }
+    if ( $request->method eq 'POST' ) {
+        if ( $request->param('device_mac') ) {
+            # User is authenticated and requesting to register a device
+            my $device_mac = clean_mac($request->param('device_mac'));
+            my $device_type;
+            $device_type = $request->param('console_type') if ( defined($request->param('console_type')) );
+            if(valid_mac($device_mac)) {
+                # Register device
+                $c->forward('registerNode', [ $pid, $device_mac, $device_type ]);
+                unless ($c->has_errors) {
+                    $c->stash(status_msg  => [ "The MAC address %s has been successfully registered.", $device_mac ]);
+                    $c->detach('landing');
+                }
             }
-        } else {
-            $c->stash(txt_auth_error => "Please verify the provided MAC address.");
         }
+    $c->stash(txt_auth_error => "Please verify the provided MAC address.");
     }
     # User is authenticated so display registration page
     $c->stash(title => "Registration", template => 'device-registration/registration.html');
